Работа над заданием 7.9
